+++ src/test/java/flab/nutridiary/product/service/ProductRegisterServiceTest.java
@@ -3,8 +3,8 @@ package flab.nutridiary.product.service;
 import flab.nutridiary.commom.generic.Nutrition;
 import flab.nutridiary.product.domain.ServingUnit;
 import flab.nutridiary.product.domain.Product;
-import flab.nutridiary.product.dto.NewProductRequest;
-import flab.nutridiary.product.dto.NewProductResponse;
+import flab.nutridiary.product.dto.request.NewProductRequest;
+import flab.nutridiary.product.dto.response.NewProductResponse;
 import flab.nutridiary.product.repository.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
